requirements.txt
pyocr
Pillow
Flask
Wand
PyPDF2